mcp-server/pyproject.toml
[project]
name = "whatsapp-mcp-server"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "requests",
    "pydantic",
    "python-dotenv",
]
